Split layoutsize from layout as they are two different things
Fix #2793
