+++ src/compile/concat.ts
@@ -3,7 +3,7 @@ import {ConcatSpec, isVConcatSpec} from '../spec';
 import {VgLayout} from '../vega.schema';
 import {BaseConcatModel} from './baseconcat';
 import {buildModel} from './common';
-import {parseConcatLayoutSize} from './layout/parse';
+import {parseConcatLayoutSize} from './layoutsize/parse';
 import {Model} from './model';
 import {RepeaterValue} from './repeater';
