+++ src/compile/layer.ts
@@ -5,8 +5,8 @@ import {flatten, keys} from '../util';
 import {VgData, VgLayout, VgScale, VgSignal, VgTitle} from '../vega.schema';
 import {parseLayerAxis} from './axis/parse';
 import {parseData} from './data/parse';
-import {assembleLayoutSignals} from './layout/assemble';
-import {parseLayerLayoutSize} from './layout/parse';
+import {assembleLayoutSignals} from './layoutsize/assemble';
+import {parseLayerLayoutSize} from './layoutsize/parse';
 import {Model} from './model';
 import {RepeaterValue} from './repeater';
 import {assembleScaleForModelAndChildren} from './scale/assemble';
