+++ src/compile/layoutsize/component.ts
