+++ src/compile/model.ts
@@ -29,9 +29,9 @@ import {ConcatModel} from './concat';
 import {DataComponent} from './data/index';
 import {FacetModel} from './facet';
 import {LayerModel} from './layer';
-import {sizeExpr} from './layout/assemble';
-import {LayoutSizeComponent, LayoutSizeIndex} from './layout/component';
 import {getHeaderGroup, getTitleGroup, HEADER_CHANNELS, HEADER_TYPES, LayoutHeaderComponent} from './layout/header';
+import {sizeExpr} from './layoutsize/assemble';
+import {LayoutSizeComponent, LayoutSizeIndex} from './layoutsize/component';
 import {assembleLegends} from './legend/assemble';
 import {LegendComponentIndex} from './legend/component';
 import {parseLegend} from './legend/parse';
