+++ test/compile/layoutsize/assemble.test.ts
@@ -4,7 +4,7 @@ import {assert} from 'chai';
 import {parseFacetModel, parseUnitModelWithScaleAndLayoutSize} from '../../util';
 
 import {X, Y} from '../../../src/channel';
-import {sizeSignals} from '../../../src/compile/layout/assemble';
+import {sizeSignals} from '../../../src/compile/layoutsize/assemble';
 import * as log from '../../../src/log';
 
 describe('compile/layout', () => {
